Create setCanvas() function
This function's purpose is to programmatically populate the canvas with grid child elements. By default it will create a 5x5 grid of 25 elements, but can be told to make more or less. Later I would like to use this with some kind of UI element to allow the user to choose the scale of the canvas.